refactor(wave4-step2): split lockfile/cache behind stable facades (#340)
* refactor(wave4-step2): add compile-safe lockfile/cache split scaffolds

* refactor(wave4-step2): move lockfile/cache logic behind stable facades

* docs(wave4-step2): add move-map, reviewer gates, and review pack

* refactor(wave4-step2): tighten facade thinness for lockfile io and cache put

* fix(wave4-step2): remove cache put facade helper coupling
