+++ crates/assay-registry/src/cache_next/errors.rs
@@ -1,0 +1,5 @@
+//! Cache error mapping scaffold.
+//!
+//! Planned ownership (Step2+):
+//! - typed error constructors/mappers only
+//! - no IO side effects
