+++ crates/assay-registry/src/cache_next/tests.rs
@@ -1,0 +1,4 @@
+//! Step2 placeholder.
+//!
+//! No tests are moved in Wave4 Step2 Commit A.
+//! Step1 freeze tests remain in existing lockfile/cache files.
