+++ crates/assay-registry/src/lockfile_next/errors.rs
@@ -1,0 +1,5 @@
+//! Lockfile error mapping scaffold.
+//!
+//! Planned ownership (Step2+):
+//! - typed error constructors/mappers only
+//! - no parsing/IO side effects
