+++ crates/assay-registry/src/lockfile_next/types.rs
@@ -1,0 +1,5 @@
+//! Lockfile type ownership scaffold.
+//!
+//! Planned ownership (Step2+):
+//! - lockfile structs/enums currently defined in lockfile facade
+//! - re-export strategy remains facade-first
